Remove duplicate merge_opens option in compile.py
The merge_opens option is available in the options received from the
command line and will apply to a Program instance and its tapes.

The merge_opens attribute is preserved on the Tape class as this allows
turning off merging (setting merge_opens to False) for specific tapes.

See #121 for more details.